+++ packages/ag_flow_cli/bricks/collection_module/__brick__/pages/{{module_file_base}}_page.dart
@@ -1,4 +1,9 @@
 import 'package:ag_flow/ag_flow.dart';
+import 'package:{{app_package_name}}/{{root_segment}}/components/{{component_namespace}}/{{component_file_base}}_appbar.dart';
+import 'package:{{app_package_name}}/{{root_segment}}/components/{{component_namespace}}/{{component_file_base}}_empty.dart';
+import 'package:{{app_package_name}}/{{root_segment}}/components/{{component_namespace}}/{{component_file_base}}_error.dart';
+import 'package:{{app_package_name}}/{{root_segment}}/components/{{component_namespace}}/{{component_file_base}}_list.dart';
+import 'package:{{app_package_name}}/{{root_segment}}/components/{{component_namespace}}/{{component_file_base}}_loading.dart';
 import 'package:{{app_package_name}}/{{root_segment}}/controllers/{{module_file_base}}_controller.dart';
 import 'package:flutter/material.dart';
 
@@ -6,13 +11,25 @@ class {{module_class_prefix}}Page extends AgBasePage<{{module_class_prefix}}Cont
   const {{module_class_prefix}}Page({super.key});
 
   @override
+  PreferredSizeWidget? appBar(BuildContext context) =>
+      {{component_class_prefix}}AppBar();
+
+  @override
+  WidgetBuilder? get loadingBuilder =>
+      (context) => const {{component_class_prefix}}Loading();
+
+  @override
+  Widget Function(BuildContext, Object, StackTrace?, VoidCallback)?
+  get errorBuilder =>
+      (context, error, stackTrace, retry) =>
+          {{component_class_prefix}}Error(error: error, onRetry: retry);
+
+  @override
+  WidgetBuilder? get emptyBuilder =>
+      (context) => const {{component_class_prefix}}Empty();
+
+  @override
   Widget buildSuccess(BuildContext context) {
-    return AgListBuilder<dynamic, int>(
-      controller: controller,
-      itemBuilder: (context, item, index) {
-        // TODO: replace with the real item widget.
-        return ListTile(title: Text(item.toString()));
-      },
-    );
+    return {{component_class_prefix}}List(controller: controller);
   }
 }