+++ packages/ag_flow_cli/test/src/commands/module_command_test.dart
@@ -107,7 +107,7 @@ void main() {
         ),
       ).readAsStringSync();
       expect(controller, contains('Future<void> add('));
-      expect(controller, contains('Future<void> update('));
+      expect(controller, contains('Future<void> updateItem('));
       expect(controller, contains('Future<void> delete('));
     },
   );
@@ -128,7 +128,7 @@ void main() {
         ),
       ).readAsStringSync();
       expect(controller, contains('Future<void> add('));
-      expect(controller, isNot(contains('Future<void> update(')));
+      expect(controller, isNot(contains('Future<void> updateItem(')));
       expect(controller, contains('Future<void> delete('));
     },
   );
@@ -149,7 +149,7 @@ void main() {
         ),
       ).readAsStringSync();
       expect(controller, isNot(contains('Future<void> add(')));
-      expect(controller, isNot(contains('Future<void> update(')));
+      expect(controller, isNot(contains('Future<void> updateItem(')));
       expect(controller, isNot(contains('Future<void> delete(')));
       expect(controller, contains('fetchPage'));
     },